Make cell quality a data object filter
